requirements.txt
Flask==3.0.2
Flask-Mail==0.9.1
Werkzeug==3.0.1
pytz==2024.1
Pillow==10.2.0
gunicorn==21.2.0
python-dotenv==1.0.1